+++ test/smokes/code_sniffer/phpcs3/expectations.rb
@@ -14,7 +14,7 @@ s.add_test(
       git_blame_info: nil
     }
   ],
-  analyzer: { name: "PHP_CodeSniffer", version: "3.5.4" }
+  analyzer: { name: "PHP_CodeSniffer", version: "3.5.5" }
 )
 
 s.add_test(
@@ -31,7 +31,7 @@ s.add_test(
       git_blame_info: nil
     }
   ],
-  analyzer: { name: "PHP_CodeSniffer", version: "3.5.4" }
+  analyzer: { name: "PHP_CodeSniffer", version: "3.5.5" }
 )
 
 s.add_test(
@@ -57,7 +57,7 @@ s.add_test(
       git_blame_info: nil
     }
   ],
-  analyzer: { name: "PHP_CodeSniffer", version: "3.5.4" }
+  analyzer: { name: "PHP_CodeSniffer", version: "3.5.5" }
 )
 
 s.add_test(
@@ -110,7 +110,7 @@ s.add_test(
       git_blame_info: nil
     }
   ],
-  analyzer: { name: "PHP_CodeSniffer", version: "3.5.4" }
+  analyzer: { name: "PHP_CodeSniffer", version: "3.5.5" }
 )
 
 s.add_test(
@@ -137,7 +137,7 @@ s.add_test(
       git_blame_info: nil
     }
   ],
-  analyzer: { name: "PHP_CodeSniffer", version: "3.5.4" }
+  analyzer: { name: "PHP_CodeSniffer", version: "3.5.5" }
 )
 
 s.add_test(
@@ -181,10 +181,10 @@ s.add_test(
       git_blame_info: nil
     }
   ],
-  analyzer: { name: "PHP_CodeSniffer", version: "3.5.4" }
+  analyzer: { name: "PHP_CodeSniffer", version: "3.5.5" }
 )
 
 s.add_test(
   "phpcs3/with_php_version",
-  type: "success", issues: :_, analyzer: { name: "PHP_CodeSniffer", version: "3.5.4" }
+  type: "success", issues: :_, analyzer: { name: "PHP_CodeSniffer", version: "3.5.5" }
 )